Extract createSystemTable to utility class
